+++ index.ts
@@ -1,3 +1,3 @@
-export * from "./dmart.model";
-export * from "./dmart.service";
+export * from "./dmart.model.js";
+export * from "./dmart.service.js";